requirements.txt
Flask==2.3.3
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
joblib==1.3.2
python-dotenv==1.0.0
